Added move method but it's not operational due to a bug in the xpos and ypos of ship. I think I made an extra object somewhere
