+++ Controller/game.h
@@ -6,8 +6,8 @@
 #define PLANET_EXPRESS_GAME_H
 
 
-#include "Helper/IOhandler.h"
-#include "player.h"
+#include "../Helper/IOhandler.h"
+#include "../Model/player.h"
 #include "headquarters.h"
 
 class game {
@@ -15,6 +15,8 @@ public:
     explicit game(IOhandler& handler);
     void run_game();
 private:
+    void handle_command(const int command_num);
+
     IOhandler handler;
     player player1;
     headquarters hq;
