+++ Model/sector.cpp
