+++ Model/sector_tile.cpp
