+++ Model/sector_tile.h
